+++ package/Runtime/DeviceSensors/Scripts/Samples.Sensors.codegen/SensorAccess.cs
@@ -8,6 +8,7 @@ namespace Needle.Typescript.GeneratedComponents
 	public partial class SensorAccess : UnityEngine.MonoBehaviour
 	{
 		public float @frequency = 60f;
+		public bool @invert = false;
 		public void start(){}
 	}
 }