+++ package/Runtime/DeviceSensors/Scripts/Samples.Sensors~/codegen/register_types.js
@@ -1,7 +1,0 @@
-﻿import { TypeStore } from "@needle-tools/engine"
-
-// Import types
-import { SensorAccess } from "../SensorAccess";
-
-// Register types
-TypeStore.add("SensorAccess", SensorAccess);