+++ package/Runtime/Lightmaps/Runtime/Module/MultiLightmap~/Lightmaps.ts
@@ -1,4 +1,4 @@
-import { Context, ContextRegistry, Interactable, SyncedTransform, delay } from "@needle-tools/engine";
+import { Context, ContextRegistry, Interactable, SyncedTransform, delay, syncField } from "@needle-tools/engine";
 import { Behaviour, GameObject } from "@needle-tools/engine";
 import { Renderer } from "@needle-tools/engine";
 import { IPointerClickHandler } from "@needle-tools/engine";
@@ -50,14 +50,10 @@ class LightmapSettings {
 //@dont-generate-component
 export class LightmapConfigurations extends Behaviour {
 
-    switchLight() {
-        this._didSwitchLightTime = Date.now();
-        this.index = this.currentIndex + 1;
-        this.context.connection.send("lightmap_index", this.currentIndex);
-    }
-
     private _didSwitchLightTime?: number;
 
+    cycleLigthmaps: boolean = true;
+    cycleInterval: number = 1;
     pingPong: boolean = false;
 
     //@type System.Collections.Generic.List<UnityEngine.Texture2D>
@@ -70,25 +66,43 @@ export class LightmapConfigurations extends Behaviour {
     set index(val: number) {
         this.setLightmap(val);
     }
-    get currentIndex(): number { return this._currentIndex; }
+    get index(): number { return this._index; }
+    get currentIndex(): number { return this.index; }
 
     private _intensities: { [key: string]: number } = {};
     private _forward: boolean = true;
     private _renderers: Renderer[] = [];
 
     awake() {
-        this.context.connection.beginListen("lightmap_index", (index: number) => {
-            this._didSwitchLightTime = Date.now()
-            this.setLightmap(index);
-        });
+        this.context.connection.beginListen("lightmap_index", this.onRemoteLightmapIndex);
+        this.context.connection.beginListen("lightmap_cycle", this.onRemoteLightmapCycle);
         this.context.connection.beginListen(RoomEvents.JoinedRoom, () => {
-            this.context.connection.send("lightmap_sync", { index: this.currentIndex, switchTime: this._didSwitchLightTime });
+            this.context.connection.send("lightmap_sync", { index: this.index, switchTime: this._didSwitchLightTime });
         });
         this.context.connection.beginListen("lightmap_sync", (model: { index: number, switchTime: number }) => {
             this.index = model.index;
             this._didSwitchLightTime = model.switchTime;
         });
     }
+
+    onDestroy() {
+        this.context.connection.stopListen("lightmap_index", this.onRemoteLightmapIndex);
+        this.context.connection.stopListen("lightmap_cycle", this.onRemoteLightmapCycle);
+    }
+
+    private onRemoteLightmapIndex = (index: number) => {
+        if(typeof index != "number") return;
+
+        this.setLightmap(index);
+        this._didSwitchLightTime = Date.now()
+    };
+
+    private onRemoteLightmapCycle = (index: number) => {
+        if(typeof index != "number") return;
+
+        this._didSwitchLightTime = undefined;
+        this.setLightmap(index);
+    };
 
     onEnable() {
         this.startCoroutine(this.switchLightmaps());
@@ -120,17 +134,35 @@ export class LightmapConfigurations extends Behaviour {
 
                 this.setLightmap(i);
             }
-            yield WaitForSeconds(1);//.3 + Math.random() * .5);
+            yield WaitForSeconds(this.cycleInterval);//.3 + Math.random() * .5);
         }
     }
 
-    private _currentIndex: number = -1;
+    private _index: number = -1;
+
+    switchLight() { this.nextLightmap(); }
+    nextLightmap() {
+        this._didSwitchLightTime = Date.now();
+        this.index = this.index + 1;
+        this.context.connection.send("lightmap_index", this.index);
+    }
+    previousLightmap() {
+        this._didSwitchLightTime = Date.now();
+        this.index = this.index + 1;
+        this.context.connection.send("lightmap_index", this.index);
+    }
+
+    selectLightmap(index: number) {
+        this.setLightmap(index);
+        this._didSwitchLightTime = Date.now();
+        this.context.connection.send("lightmap_index", this.index);
+    }
 
     private setLightmap(index: number) {
         if (typeof index !== "number") return;
         if (!this.lightmaps) return;
         index %= this.lightmaps.length;
-        this._currentIndex = index;
+        this._index = index;
         for (const rend of this._renderers) {
             rend.lightmap = this.lightmaps[index];
         }
@@ -149,6 +181,12 @@ export class LightmapConfigurations extends Behaviour {
                 }
             }
         }
+    }
+
+    startCyclingLightmaps() {
+        this._didSwitchLightTime = undefined;
+        this.setLightmap(this.index + 1);
+        this.context.connection.send("lightmap_cycle", this.index);
     }
 
     private disableRendererEmission(renderer: Renderer) {