+++ package/Runtime/Scrollytelling/Needle.Samples.Scrollytelling.codegen/ScrollTimelineSlider.cs
@@ -1,0 +1,16 @@
+// NEEDLE_CODEGEN_START
+// auto generated code - do not edit directly
+
+#pragma warning disable
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public partial class ScrollTimelineSlider : UnityEngine.MonoBehaviour
+	{
+		public UnityEngine.Playables.PlayableDirector @timeline;
+		public void OnEnable(){}
+		public void OnDisable(){}
+	}
+}
+
+// NEEDLE_CODEGEN_END