+++ package/Runtime/ShootingRange/Scripts/ShootingRange~/Gun.ts
@@ -1,4 +1,4 @@
-import { Animator, AudioSource, Behaviour, EventList, GameObject, Gizmos, ParticleSystem, Watch, WebXR, getParam, isMobileDevice, serializable, setWorldPosition, setWorldQuaternion } from "@needle-tools/engine";
+import { Animator, AudioSource, Behaviour, EventList, GameObject, Gizmos, NeedleXREventArgs, NeedleXRSession, ParticleSystem, Watch, WebXR, getParam, getTempVector, isMobileDevice, serializable, setWorldPosition, setWorldQuaternion } from "@needle-tools/engine";
 
 import { Object3D, Quaternion, Vector2, Vector3 } from "three";
 import { Target } from "./Target";
@@ -6,7 +6,31 @@ import { Target } from "./Target";
 
 const debug = getParam("debugfps");
 
-class GunEffects {
+
+// gathers input, performs raycast, controls effects and handles hitting a target
+export class Gun extends Behaviour {
+
+    @serializable()
+    enableMobileInput: boolean = true;
+
+    @serializable()
+    enableDesktopInput: boolean = true;
+
+    //@nonSerialized
+    @serializable()
+    vrSide: XRHandedness | "any" = "any";
+
+    @serializable()
+    fireRate: number = 0.1;
+
+    @serializable(Object3D)
+    raycastReference?: Object3D;
+
+    @serializable(Animator)
+    gunAnimator?: Animator;
+
+    @serializable()
+    fireAnimation: string = "Fire";
 
     @serializable(AudioSource)
     fireSound?: AudioSource;
@@ -20,184 +44,20 @@ class GunEffects {
     @serializable(ParticleSystem)
     impactEffect?: ParticleSystem;
 
-    @serializable()
-    impactOffset: number = 0.3;
-}
-
-class GunStats {
-
-    @serializable()
-    fireRate: number = 0.1;
-}
-
-class GunReferences {
-
-    // refrences - physics
-    @serializable(Object3D)
-    raycastReference?: Object3D;
-
-    @serializable()
-    scaleInVR: number = 0.1;
-}
-
-class GunAnimation {
-
-    @serializable(Animator)
-    gunAnimator?: Animator;
-
-    @serializable()
-    fireAnimation: string = "Fire";
-}
-
-
-export enum GunInputEnum {
-    Left = 0,
-    Solo = 0,
-    Right = 2,
-}
-
-// gathers input, performs raycast, controls effects and handles hitting a target
-export class Gun extends Behaviour {
-
-    @serializable(GunEffects)
-    effects!: GunEffects;
-
-    @serializable(GunStats)
-    stats!: GunStats;
-
-    @serializable(GunReferences)
-    references!: GunReferences;
-
-    @serializable(GunAnimation)
-    animation!: GunAnimation;
-
-    @serializable()
-    vrHideControllers: boolean = true;
-
-    @serializable()
-    vrHideHands: boolean = true;
-
-    @serializable()
-    enableOnlyRightWeaponOnMobile: boolean = true;
-
-    @serializable()
-    enableMobileInput: boolean = true;
-
-    @serializable()
-    enableDesktopInput: boolean = true;
-
-    // input
-    @serializable()
-    gunInput: GunInputEnum = 0;
+    // --------------------
 
     // reporting events 
-    @serializable(EventList)
-    onHitTarget!: EventList;
+    static onHitTarget: EventList = new EventList();
+    static onMiss: EventList = new EventList();
 
-    @serializable(EventList)
-    onMiss!: EventList;
-
-    private lastPosition = new Vector3(0, 0, 0);
-    private currentPosition = new Vector3(0, 0, 0);
-
-    private characterGunPosition = new Vector3(0, 0, 0);
-    private characterGunRotation = new Quaternion(0, 0, 0, 1);
-
-    // raycast
-    private raycastWorldDirection = new Vector3();
-    private raycastWorldOrigin = new Vector3();
-    private isVR: boolean = false;
-
-    private parentOnStart?: Object3D;
-    private webXR?: WebXR;
-
-    awake() {
-        this.gameObject.getWorldPosition(this.lastPosition);
-        this.gameObject.getWorldPosition(this.currentPosition);
-
-        this.characterGunPosition.copy(this.gameObject.position);
-        this.characterGunRotation.copy(this.gameObject.quaternion);
-
-        this.parentOnStart = this.gameObject.parent!;
-
-        this.webXR = GameObject.findObjectOfType(WebXR)!;
-    }
-
-    update(): void {
-        if (this.webXR && this.isVR !== this.webXR.IsInVR) {
-            this.isVR = this.webXR.IsInVR;
-            this.onVRChanged(this.isVR);
-        }
-    }
-    onBeforeRender(): void {
-
-        if (this.references.raycastReference) {
-
-            this.references.raycastReference.getWorldPosition(this.raycastWorldOrigin);
-            this.references.raycastReference.getWorldDirection(this.raycastWorldDirection);
-        }
-
-        let wPos = new Vector3();
-        this.gameObject.getWorldPosition(wPos);
-
-        if (this.webXR) {
-
-            const isLeftHand = this.gunInput == GunInputEnum.Left;
-            const controller = isLeftHand ? this.webXR.LeftController : this.webXR.RightController;
-
-            if (this.isVR && controller != null) {
-
-                controller.controller.getWorldPosition(this.gameObject.position);
-                this.gameObject.quaternion.copy(controller.rayRotation);
-                
-                this.gameObject.rotateY(Math.PI); // ugly, but FWD is inverted
-                
-                if(controller.hand.visible) {
-                    this.gameObject.rotateZ(Math.PI); 
-                }
-
-                if (controller.isUsingHands) {
-                    const negate = isLeftHand ? 1 : -1;
-                    this.gameObject.rotateZ(Math.PI * 0.5 * negate);
-                }
-
-                if (controller.selectionDown) {
+    onUpdateXR(args: NeedleXREventArgs): void {
+        args.xr.controllers.forEach(c => {
+            if (c.getButton("primary")?.isDown === true) {
+                if (c.side == this.vrSide || this.vrSide === "any") {
                     this.fire();
                 }
             }
-        }
-    }
-
-    onVRChanged(isVR: boolean) {
-
-        if (!this.webXR)
-            return;
-
-        this.gameObject.parent?.remove(this.gameObject);
-        if (isVR) //enter VR
-        {
-            this.webXR.Controllers.forEach(c => {
-
-                c.showRaycastLine = false;
-
-                if (this.vrHideControllers)
-                    c.controllerModel.visible = false;
-
-                if (this.vrHideHands)
-                    c.handPointerModel.visible = false;
-            });
-
-            this.context.scene.add(this.gameObject);
-            const s = this.references.scaleInVR;
-            this.gameObject.scale.set(s, s, s);
-        }
-        else //exit VR
-        {
-            this.parentOnStart?.add(this.gameObject);
-            this.gameObject.position.copy(this.characterGunPosition);
-            this.gameObject.quaternion.copy(this.characterGunRotation);
-            this.gameObject.scale.set(1, 1, 1);
-        }
+        });
     }
 
     // subrscribe to input events
@@ -229,21 +89,18 @@ export class Gun extends Behaviour {
         this._lastTouchEndPoint.set(x, y);
         const dist = Math.sqrt(dx * dx + dy * dy);
 
-        console.log("touch end", event);
         if (dist < 15) {
             this.fire();
-            console.log("tap finger");
         }
         if (event.touches.length <= 0) { // last finger 
             this.fire(true, true);
-            console.log("last finger");
         }
     }
 
     private onMouseClick = (event: MouseEvent) => {
         if(isMobileDevice()) return; // ignore desktop input on mobile
 
-        if (event.button != this.gunInput) // if not the correct mouse button, abort
+        if (event.button !== 0) // if not the LMB or primary, abort
             return;
         this.fire();
     }
@@ -260,8 +117,9 @@ export class Gun extends Behaviour {
 
     fire(applyFireRate: boolean = true, ignoreMiss: boolean = false) {
         const t = this.context.time.time;
-        if (applyFireRate && t < this.fireTimeStamp + this.stats.fireRate) // fire rate
+        if (applyFireRate && t < this.fireTimeStamp + this.fireRate) // fire rate
             return;
+
         this.fireTimeStamp = t; // save the time of a successful fire
         const hit = this.firePhysically(ignoreMiss);
         if(ignoreMiss && !hit) return; // abort miss
@@ -270,32 +128,38 @@ export class Gun extends Behaviour {
 
     fireVisually(hitPoint: Vector3 | null) {
         // fire anim
-        this.animation.gunAnimator?.play(this.animation.fireAnimation);
+        this.gunAnimator?.play(this.fireAnimation);
         // particles
-        this.effects.muzzleFlash?.play();
-        this.effects.ejectShell?.play();
+        this.muzzleFlash?.play();
+        this.ejectShell?.play();
         // audio
-        this.effects.fireSound?.stop();
-        this.effects.fireSound?.play();
+        this.fireSound?.stop();
+        this.fireSound?.play();
 
         // Setup hit particle effect: 
-        if (this.effects.impactEffect && this.raycastWorldOrigin && hitPoint) {
-            setWorldPosition(this.effects.impactEffect.gameObject, hitPoint);
+        if (this.impactEffect && hitPoint) {
+            setWorldPosition(this.impactEffect.gameObject, hitPoint);
 
             // play the effect
-            this.effects.impactEffect.stop();
-            this.effects.impactEffect.play();
+            this.impactEffect.stop();
+            this.impactEffect.play();
         }
     }
 
     // perform raycast and reports hit/miss
     firePhysically(ignoreMiss: boolean = false): Vector3 | null {
-        if (!this.raycastWorldOrigin)
+        if (!this.raycastReference)
             return null;
 
         const physics = this.context.physics;
+        const shootRef = this.raycastReference;
 
-        const hit = physics.raycastPhysicsFast(this.raycastWorldOrigin, this.raycastWorldDirection);
+        const hit = physics.engine?.raycast(shootRef.getWorldPosition(getTempVector()), shootRef.getWorldDirection(getTempVector()));
+        if (debug) {
+            const from = shootRef.getWorldPosition(getTempVector());
+            const to = hit ? hit.point : from.add(shootRef.getWorldDirection(getTempVector()).multiplyScalar(10));
+            Gizmos.DrawLine(from, to, 0xff0000, .25, false);
+        }
         if (hit) {
             if (debug) {
                 Gizmos.DrawWireSphere(hit.point, 0.1, 0xff0000, .1);
@@ -306,7 +170,7 @@ export class Gun extends Behaviour {
                 if (!target.isShot) {
                     // only perform hit once
                     target.performHit();
-                    this.onHitTarget?.invoke(this, target);
+                    Gun.onHitTarget?.invoke(this, target);
                     return hit.point;
                 }
             }
@@ -316,7 +180,7 @@ export class Gun extends Behaviour {
         }
 
         if (!ignoreMiss) { // miss            
-            this.onMiss?.invoke(this);
+            Gun.onMiss?.invoke(this);
         }
 
         return null;