+++ package/Runtime/_Shared/SampleScripts.codegen/Cannon.cs
@@ -10,8 +10,10 @@ namespace Needle.Typescript.GeneratedComponents
 		public UnityEngine.GameObject @prefab;
 		public UnityEngine.AudioSource @audioSource;
 		public float @strength = 10f;
+		public float @maxInstances = 10f;
 		public void start(){}
-		public void update(){}
+		public void OnEnable(){}
+		public void OnDisable(){}
 	}
 }