+++ package/Runtime/_Shared/SampleScripts.codegen/SimpleLeaderboard.cs
@@ -1,0 +1,17 @@
+// NEEDLE_CODEGEN_START
+// auto generated code - do not edit directly
+
+#pragma warning disable
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public partial class SimpleLeaderboard : UnityEngine.MonoBehaviour
+	{
+		// public LeaderboardModel @data; → Could not resolve C# type
+		public void OnEnable(){}
+		public void insertScore(string @name, float @score){}
+		public void onDataChanged(object @evt){}
+	}
+}
+
+// NEEDLE_CODEGEN_END