+++ package/Runtime/_Shared/SampleScripts.codegen/XRLifecycleEvents.cs
@@ -9,8 +9,8 @@ namespace Needle.Typescript.GeneratedComponents
 	{
 		public UnityEngine.Events.UnityEvent @onSessionStart = new UnityEngine.Events.UnityEvent();
 		public UnityEngine.Events.UnityEvent @onSessionEnd = new UnityEngine.Events.UnityEvent();
-		public void OnEnable(){}
-		public void OnDisable(){}
+		public void onEnterXR(object @_args){}
+		public void onExitXR(object @_args){}
 	}
 }