+++ package/Runtime/_Shared/SampleScripts~/EmitParticlesOnClick.ts
@@ -1,5 +1,5 @@
 // START MARKER Emit Particles On Click
-import { InputEvents, setWorldPosition, Behaviour, GameObject, Gizmos, ParticleSystem, getParam, serializeable, showBalloonMessage, showBalloonWarning, WebXR, getWorldPosition } from "@needle-tools/engine";
+import { setWorldPosition, Behaviour, GameObject, Gizmos, ParticleSystem, getParam, serializeable, NeedleXREventArgs } from "@needle-tools/engine";
 import { Vector3 } from "three";
 
 const debug = getParam("sample_debugParticlesOnClick");
@@ -9,28 +9,23 @@ export class EmitParticlesOnClick extends Behaviour {
     @serializeable(ParticleSystem)
     particleSystems: ParticleSystem[] = [];
 
-    private webXR?: WebXR;
-    private tempVector: Vector3 = new Vector3();
-
-    awake(): void {
-        this.webXR = GameObject.findObjectOfType(WebXR)!;
-    }
-
-    update() {
+    update(): void {
         if(!this.context.isInVR && this.context.input.getPointerClicked(0)) {
             const hits = this.context.physics.raycast();
             if(hits.length > 0) {
                 this.spawnParticlesAt(hits[0].point);
             }
-        }
-        else if(this.context.isInVR && this.webXR) {
-            this.webXR.Controllers.forEach((controller) => {
-                if (controller.selectionDown && controller.raycastHitPoint) {
-                    getWorldPosition(controller.raycastHitPoint, this.tempVector);
-                    this.spawnParticlesAt(this.tempVector);
+        }  
+    }
+    onUpdateXR(args: NeedleXREventArgs) {
+        args.xr.controllers.forEach((controller) => {
+            if (controller.getButton("primary")?.isDown === true) {
+                const hit = this.context.physics.raycastFromRay(controller.ray)?.at(0);
+                if(hit) {
+                    this.spawnParticlesAt(hit.point);   
                 }
-            })
-        }
+            }
+        });
     }
 
     spawnParticlesAt(pos: Vector3) {