+++ package/Runtime/_Shared/SampleScripts~/PlayAnimationOnCollision.ts
@@ -9,7 +9,7 @@ export class PlayAnimationOnCollision extends Behaviour {
 
     private _lastTriggerTime: number = 0;
 
-    onTriggerEnter(col) {
+    onTriggerEnter(_col) {
         if (this.context.time.time - this._lastTriggerTime < .3) return;
         this._lastTriggerTime = this.context.time.time;
         if (this.animation) this.animation.play(0, { loop: false });