+++ package/WebProjects~/VueTemplate/src/main.ts
@@ -1,5 +1,0 @@
-import { createApp } from 'vue'
-//@ts-ignore
-import App from './App.vue'
-const app = createApp(App);
-app.mount('#app');